Refactor file utilities and add module for large-file detection
