+++ dreamos/core/agent_loop.py
@@ -17,9 +17,11 @@ from datetime import datetime
 from .utils.file_utils import (
     safe_read,
     safe_write,
+    ensure_dir,
+)
+from .utils.json_utils import (
     load_json,
     save_json,
-    ensure_dir
 )
 from social.utils.log_manager import LogManager, LogConfig, LogLevel
 from .agent_control.controller import AgentController
