+++ dreamos/core/autonomy_loop_runner.py
@@ -19,9 +19,11 @@ import logging
 from .utils.file_utils import (
     safe_read,
     safe_write,
+    ensure_dir,
+)
+from .utils.json_utils import (
     load_json,
     save_json,
-    ensure_dir
 )
 from social.utils.log_manager import LogManager, LogConfig, LogLevel
