+++ dreamos/core/metrics.py
@@ -10,10 +10,10 @@ from dreamos.core.utils.file_utils import (
     ensure_dir,
     safe_write,
     safe_read,
-    read_json,
-    write_json,
+)
+from dreamos.core.utils.json_utils import (
     load_json,
-    save_json
+    save_json,
 )
 
 class CommandMetrics:
