+++ dreamos/core/persistent_queue.py
@@ -21,8 +21,10 @@ from .utils.file_utils import (
     ensure_dir,
     safe_write,
     safe_read,
+)
+from .utils.json_utils import (
     load_json,
-    save_json
+    save_json,
 )
 from social.utils.log_manager import LogManager
