+++ dreamos/core/response_collector.py
@@ -25,9 +25,11 @@ import sys
 from dreamos.core.utils.file_utils import (
     safe_read,
     safe_write,
+    ensure_dir,
+)
+from dreamos.core.utils.json_utils import (
     load_json,
     save_json,
-    ensure_dir
 )
 from social.utils.log_manager import LogManager
 from social.utils.log_config import LogConfig
