+++ dreamos/core/utils/file_utils.py
@@ -5,8 +5,6 @@ Robust file system operations with Windows compatibility.
 """
 
 import os
-import json
-import yaml
 import logging
 import tempfile
 import shutil
@@ -135,76 +133,6 @@ def safe_read(file_path: Union[str, Path], encoding: str = 'utf-8') -> Optional[
         logger.error(f"Failed to read file {file_path}: {e}")
         return None
 
-def load_json(file_path: Union[str, Path]) -> Optional[Dict[str, Any]]:
-    """Load JSON from file.
-    
-    Args:
-        file_path: Path to JSON file
-        
-    Returns:
-        Parsed JSON or None if failed
-    """
-    content = safe_read(file_path)
-    if content is None:
-        return None
-    try:
-        return json.loads(content)
-    except json.JSONDecodeError as e:
-        logger.error(f"Failed to parse JSON from {file_path}: {e}")
-        return None
-
-def save_json(file_path: Union[str, Path], data: Dict[str, Any], indent: int = 2) -> bool:
-    """Save data as JSON.
-    
-    Args:
-        file_path: Path to save to
-        data: Data to save
-        indent: JSON indentation
-        
-    Returns:
-        True if successful
-    """
-    try:
-        content = json.dumps(data, indent=indent)
-        return safe_write(file_path, content)
-    except Exception as e:
-        logger.error(f"Failed to save JSON to {file_path}: {e}")
-        return False
-
-def load_yaml(file_path: Union[str, Path]) -> Optional[Dict[str, Any]]:
-    """Load YAML from file.
-    
-    Args:
-        file_path: Path to YAML file
-        
-    Returns:
-        Parsed YAML or None if failed
-    """
-    content = safe_read(file_path)
-    if content is None:
-        return None
-    try:
-        return yaml.safe_load(content)
-    except yaml.YAMLError as e:
-        logger.error(f"Failed to parse YAML from {file_path}: {e}")
-        return None
-
-def save_yaml(file_path: Union[str, Path], data: Dict[str, Any]) -> bool:
-    """Save data as YAML.
-    
-    Args:
-        file_path: Path to save to
-        data: Data to save
-        
-    Returns:
-        True if successful
-    """
-    try:
-        content = yaml.dump(data, default_flow_style=False)
-        return safe_write(file_path, content)
-    except Exception as e:
-        logger.error(f"Failed to save YAML to {file_path}: {e}")
-        return False
 
 def ensure_dir(dir_path: Union[str, Path]) -> bool:
     """Ensure directory exists.
@@ -262,108 +190,6 @@ def find_files(dir_path: Union[str, Path], pattern: str) -> List[Path]:
         logger.error(f"Failed to find files in {dir_path}: {e}")
         return []
 
-def read_json(path: Union[str, Path], default: Any = None) -> Dict[str, Any]:
-    """Read JSON from a file.
-    
-    Args:
-        path: File path to read from
-        default: Default value to return if file doesn't exist or is invalid
-        
-    Returns:
-        Parsed JSON data or default value
-        
-    Raises:
-        FileOpsError: If file doesn't exist and no default provided
-        FileOpsIOError: If file exists but contains invalid JSON
-    """
-    path = Path(path)
-    if not path.exists():
-        if default is not None:
-            return default
-        raise FileOpsError(f"File not found: {path}")
-        
-    try:
-        with open(path) as f:
-            return json.load(f)
-    except (json.JSONDecodeError, Exception) as e:
-        if default is not None:
-            return default
-        raise FileOpsIOError(f"Invalid JSON in {path}: {e}")
-
-def write_json(path: Union[str, Path], data: dict) -> None:
-    """Write JSON to a file.
-    
-    Args:
-        path: File path to write to
-        data: Data to write
-    """
-    path = Path(path)
-    path.parent.mkdir(parents=True, exist_ok=True)
-    with open(path, 'w') as f:
-        json.dump(data, f, indent=2)
-
-def read_yaml(path: Union[str, Path], default: Any = None) -> Any:
-    """Read YAML from a file.
-    
-    Args:
-        path: File path to read from
-        default: Default value to return if file doesn't exist or is invalid
-        
-    Returns:
-        Loaded YAML data or default value
-        
-    Raises:
-        FileOpsError: If file doesn't exist and no default provided
-        FileOpsIOError: If file exists but contains invalid YAML
-    """
-    path = Path(path)
-    if not path.exists():
-        if default is not None:
-            return default
-        raise FileOpsError(f"File not found: {path}")
-        
-    try:
-        with open(path) as f:
-            content = f.read()
-            if not content.strip():  # Empty or whitespace-only file
-                return default if default is not None else {}
-            # Use strict YAML loading
-            try:
-                return yaml.safe_load(content)
-            except yaml.YAMLError as e:
-                # Double check with a more strict parser
-                try:
-                    yaml.scanner.Scanner(content)
-                    yaml.parser.Parser(content)
-                except Exception as e2:
-                    if default is not None:
-                        return default
-                    raise FileOpsIOError(f"Invalid YAML in {path}: {e2}")
-                # If we get here, the YAML is actually valid
-                return yaml.safe_load(content)
-    except Exception as e:
-        if default is not None:
-            return default
-        raise FileOpsIOError(f"Error reading YAML from {path}: {e}")
-
-def write_yaml(path: Union[str, Path], data: dict) -> None:
-    """Write YAML to a file.
-    
-    Args:
-        path: File path to write to
-        data: Data to write
-        
-    Raises:
-        FileOpsPermissionError: If file is read-only or permission denied
-        FileOpsIOError: If I/O error occurs
-    """
-    try:
-        with open(path, 'w') as f:
-            yaml.safe_dump(data, f, default_flow_style=False)
-    except PermissionError as e:
-        raise FileOpsPermissionError(f"Permission denied: {path}") from e
-    except IOError as e:
-        raise FileOpsIOError(f"I/O error: {path}") from e
 
 def safe_rmdir(path: Union[str, Path], recursive: bool = False) -> None:
     """Safely remove a directory.
